Isolate toolbox filters from rest of test config using intgration test.
