+++ lib/galaxy_test/selenium/test_personal_information.py
@@ -1,5 +1,3 @@
-from selenium.common.exceptions import NoSuchElementException
-
 from .framework import (
     selenium_test,
     SeleniumTestCase,
@@ -7,32 +5,6 @@ from .framework import (
 
 
 class ManageInformationTestCase(SeleniumTestCase):
-
-    @selenium_test
-    def test_toolbox_filters(self):
-        '''
-        Test applying and removing a toolbox filter.
-
-        This test applies the filter galaxy.selenium.filters.toolbox:restrict_test and confirms that
-        the specified section is no longer displayed in the browser.
-        '''
-        self.register()
-        # The tool panel section should be visible and clickable at this stage
-        section = self.driver.find_element_by_link_text('Test Section')
-        self.action_chains().move_to_element(section).click().perform()
-        self.navigate_to_user_preferences()
-        self.components.preferences.toolbox_filters.wait_for_and_click()
-        self.sleep_for(self.wait_types.UX_RENDER)
-        sibling_text = 'This tool filter will disable the Test Section section.'
-        # This is the least terrible way I've found to get the right element.
-        filter_upload = self.driver.find_element_by_xpath("//span[contains(. ,'%s')]/../div//label" % sibling_text)
-        self.action_chains().move_to_element(filter_upload).click().perform()
-        self.sleep_for(self.wait_types.UX_RENDER)
-        self.components.toolbox_filters.submit.wait_for_and_click()
-        self.sleep_for(self.wait_types.UX_RENDER)
-        self.home()
-        # But now it should raise NoSuchElementException
-        self.assertRaises(NoSuchElementException, lambda: self.driver.find_element_by_link_text('Test Section'))
 
     @selenium_test
     def test_api_key(self):
@@ -145,11 +117,6 @@ class ManageInformationTestCase(SeleniumTestCase):
         for input_field_label in address_fields.keys():
             input_field = self.get_address_input_field(get_address_form(), input_field_label)
             self.assertTrue(input_field.get_attribute('value') == address_fields[input_field_label])
-
-    def navigate_to_user_preferences(self):
-        self.home()
-        self.click_masthead_user()
-        self.components.masthead.preferences.wait_for_and_click()
 
     def navigate_to_manage_information(self):
         self.navigate_to_user_preferences()
